+++ server.js
@@ -1,43 +1,28 @@
-const express = require('express');
-const mongoose = require('mongoose');
-const app = express();
-const shortUrl = require('./models/shortUrl');
-const cors = require('cors') 
 
-// enabling CORS for any unknown origin(https://xyz.example.com) 
-app.use(cors()); 
+const { MongoClient, ServerApiVersion } = require('mongodb');
+const password = encodeURIComponent('%40tB0E0Z0F7');
+const uri = `mongodb+srv://AdrianLP:${password}@cluster0.c5x5l0p.mongodb.net/`;
+// const uri = `mongodb://127.0.0.1:27017/URL-S`;
 
-app.use(express.static(__dirname + '/public'));
-app.use(express.urlencoded({ extendend: true }));
-
-//DB
-mongoose.connect('mongodb://localhost:5000/urlShortener', {
-
-})
-
-app.set('view engine', 'ejs');
-
-app.get('/', async (req, res) => {
-    const shortUrls = await shortUrl.find();
-    res.render('index', { shortUrls: shortUrls });
+// Create a MongoClient with a MongoClientOptions object to set the Stable API version
+const client = new MongoClient(uri, {
+  serverApi: {
+    version: ServerApiVersion.v1,
+    strict: true,
+    deprecationErrors: true,
+  }
 });
 
-app.post('/shortUrls', async (req, res) => {
-    await shortUrl.create({
-        full: req.body.fullUrl
-    })
-    res.redirect('/');
-});
-
-app.get('/:shortUrl', async (req, res) => {
-    const shortUrl = await ShortUrl.findOne({ short: req.params.shortUrl })
-    if (shortUrl == null) return res.sendStatus(404)
-  
-    shortUrl.clicks++
-    shortUrl.save()
-  
-    res.redirect(shortUrl.full)
-  })
-  
-
-app.listen(process.env.PORT || 5000);
+async function run() {
+  try {
+    // Connect the client to the server	(optional starting in v4.7)
+    await client.connect();
+    // Send a ping to confirm a successful connection
+    await client.db("admin").command({ ping: 1 });
+    console.log("Pinged your deployment. You successfully connected to MongoDB!");
+  } finally {
+    // Ensures that the client will close when you finish/error
+    await client.close();
+  }
+}
+run().catch(console.dir);
